+++ renderer/units/JavaScript Runner/samples/Emoji/Display Emoji Groups.js
@@ -1,3 +1,3 @@
 // Display Emoji Groups
-const emojiTestGroups = require ('emoji-test-groups');
+const emojiTestGroups = require ('./lib/unicode/emoji-test-groups.js');
 $.write ($.stringify (emojiTestGroups, null, 4));