+++ renderer/units/JavaScript Runner/samples/Emoji/Display Emoji List.js
@@ -1,3 +1,3 @@
 // Display Emoji List
-const emojiTestList = require ('emoji-test-list');
+const emojiTestList = require ('./lib/unicode/emoji-test-list.js');
 $.write ($.stringify (emojiTestList, null, 4));