+++ renderer/units/JavaScript Runner/samples/Emoji/Get Component Emoji String.js
@@ -1,5 +1,5 @@
 // Get Component Emoji String
-const emojiTestList = require ('emoji-test-list');
+const emojiTestList = require ('./lib/unicode/emoji-test-list.js');
 let emojis = Object.keys (emojiTestList).filter (emoji => emojiTestList[emoji].isComponent);
 $.writeln ("Component Emoji:", emojis.length);
 $.writeln (emojis.join (""));