+++ renderer/units/JavaScript Runner/samples/Emoji/Test Exec Emoji Test Patterns.js
@@ -1,5 +1,5 @@
 // Test Exec Emoji Test Patterns
-const emojiTestPatterns = require ('emoji-test-patterns');
+const emojiTestPatterns = require ('./lib/unicode/emoji-test-patterns.js');
 const allEmojiRegex = new RegExp (emojiTestPatterns["Emoji_Test_All"], 'gu');
 const text = "AaĀā👩‍❤️‍💋‍👨#*0❤🇦愛爱애💜";
 let match;