+++ renderer/units/Unihan Data Finder/tag-data-table.js
@@ -1,8 +1,7 @@
 //
 const regexp = require ('../../lib/unicode/regexp.js');
 const unicode = require ('../../lib/unicode/unicode.js');
-const unihanData = require ('../../lib/unicode/parsed-unihan-data.js');
-const getCompatibilitySource = require ('../../lib/unicode/get-cjk-compatibility-source.js');
+const unihan = require ('../../lib/unicode/unihan.js');
 //
 const deferredSymbols = (process.platform === 'darwin');
 //
@@ -90,30 +89,7 @@ module.exports.create = function (characterInfos, params)
             let data = unicode.getCharacterBasicData (character);
             let dataRow = document.createElement ('tr');
             dataRow.className = 'data-row';
-            let status = regexp.isUnified (character) ? "Unified Ideograph" : "Compatibility Ideograph";
-            let source = (!regexp.isUnified (character)) ? getCompatibilitySource (character) : "";
-            let set = "Full Unihan";
-            let tags = unihanData.codePoints[data.codePoint];
-            if ("kIICore" in tags)
-            {
-                set = "IICore";
-            }
-            else if ("kUnihanCore2020" in tags)
-            {
-                set = "Unihan Core (2020)";
-            }
-            let lines =
-            [
-                `Code Point: ${data.codePoint}`,
-                `Age: Unicode ${data.age} (${data.ageDate})`,
-                `Set: ${set}`,
-                `Status: ${status}`
-            ];
-            if (source)
-            {
-                lines.push (`Source: ${source}`);
-            }
-            dataRow.title = lines.join ("\n");
+            dataRow.title = unihan.getTooltip (character);
             let symbolData = document.createElement ('td');
             symbolData.className = 'symbol-data';
             if (deferredSymbols)